% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input_parameters.R
\name{generate_input_parameters}
\alias{generate_input_parameters}
\title{Function to generate random parameters to the Markov model
These are based on 'parameters' defined in hips_input_data.xlsx}
\usage{
generate_input_parameters(n_samples)
}
\arguments{
\item{n_samples}{Number of samples to generate}
}
\value{
Matrix with n_samples rows and 15 columns, one for each parameter
}
\description{
Function to generate random parameters to the Markov model
These are based on 'parameters' defined in hips_input_data.xlsx
}
\examples{
sampled_input_parameters <- generate_input_parameters(10)
}
