% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_net_benefit.R
\name{generate_net_benefit}
\alias{generate_net_benefit}
\title{Generate net benefit}
\usage{
generate_net_benefit(input_parameters, lambda = 20000)
}
\arguments{
\item{input_parameters}{Matrix with row for each sample and column 
for each parameter, with values samples for the model input parameters}

\item{lambda}{Willingness-to-pay threshold for net monetary benefit}
}
\value{
List of matrices total_costs, total_qalys, net_benefit and 
incremental_net_benefit with n_implants rows and n_samples columns
}
\description{
Function to use Markov modelling to simulate total costs, QALYs and net benefit
}
\examples{
# First sample the input parameters
input_parameters <- generate_input_parameters(n_samples)

# Generate the net benefit using the above input parameters
model_output <- generate_net_benefit(input_parameters, lambda = 20000)

# Average costs
rowMeans(model_output$total_costs)

# Average QALYS
rowMeans(model_output$total_qalys)
}
