% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_state_costs.R
\name{generate_state_costs}
\alias{generate_state_costs}
\title{Generate state costs}
\usage{
generate_state_costs(input_parameters)
}
\arguments{
\item{input_parameters}{Matrix with row for each sample and column 
for each parameter, with values samples for the model input parameters}
}
\value{
Array of n_implants x n_samples x n_states, with values equal to costs 
for each state for each implant and sample
}
\description{
Function to generate state costs for each implant, sample, and state. This is
used internally by the generate_net_benefit() function
}
\examples{
# First sample the input parameters
input_parameters <- generate_input_parameters(n_samples)

# Generate the state costs
state_costs <- generate_state_costs(input_parameters)

# Sampled costs for first implant
state_costs[1, , ]
}
